Update README.md
Changed to handle code which works with guardian id to get media for multiple children. Also noted that it creates multiple media folders, one per child. I also reordered the entries in the file since the CSRF token tends to be a little long and the cookie tends to be very long (by human standards).