Update brightwheel.py
I have many additions to this code, but <GUILTY ADMISSION> I find it incredibly hard to read with two-space tabs. So I am first making a commit changing the old code to 4-space tabs -- so that the actual changes to code logic will be obvious in comparison to this version. Apologies to Michael Li Tianhui: I am sorry for making a change to esthetics; that always irked me when coauthors did that -- so I don't do it lightly here. However, I found it so hard to read without this change. Mea culpa, mea culpa, mea maxima culpa.